requirements.txt
coreapi==2.3.3
djangorestframework_jwt==1.11.0
Django==2.1.1
dj_database_url==0.4.2
coreschema==0.0.4
numpy==1.15.1
psycopg2==2.7.5
django_rest_swagger==2.2.0
django_environ==0.4.5
environ==1.0
PyJWT==1.6.4
djangorestframework==3.8.2
whitenoise==4.1
gunicorn==19.9.0
django-templateaddons3==1.0
argon2-cffi==18.3.0
pandas==0.23.0
pillow==5.2.0

